move some generic popup action to xdsmlframework.ide.ui
the following popup actions on project is now moved from
sequentiel.javaxdsml.ide.ui to xdsmlframework.ide.ui:
-
