+++ framework/xdsml_framework/plugins/org.eclipse.gemoc.xdsmlframework.ide.ui/plugin.xml
@@ -107,6 +107,99 @@
                name="xdsmlGroup">
          </separator>
       </menuContribution>
+      <menuContribution
+            allPopups="false"
+            locationURI="popup:org.eclipse.ui.popup.any?after=xdsmlGroup">
+         <menu
+               icon="icons/IconeGemocLanguage-16.png"
+               id="GemocLanguage.menu"
+               label="GEMOC Language">
+            <command
+                  commandId="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateDomainModelProject"
+                  icon="icons/plus-16.png"
+                  label="Create Domain Model Project for language"
+                  style="push">
+               <visibleWhen
+                     checkEnabled="false">
+                  <with
+                        variable="selection">
+                     <count
+                           value="1">
+                     </count>
+                     <iterate
+                           ifEmpty="false">
+                        <reference
+                              definitionId="org.eclipse.gemoc.xdsmlframework.ide.ui.definitions.IsSelectionInGemocLanguageProject">
+                        </reference>
+                     </iterate>
+                  </with>
+               </visibleWhen>
+            </command>
+            <command
+                  commandId="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateSiriusEditorProject"
+                  icon="icons/plus-16.png"
+                  label="Create Sirius Editor Project for language"
+                  style="push">
+               <visibleWhen
+                     checkEnabled="false">
+                  <with
+                        variable="selection">
+                     <count
+                           value="1">
+                     </count>
+                     <iterate
+                           ifEmpty="false">
+                        <reference
+                              definitionId="org.eclipse.gemoc.xdsmlframework.ide.ui.definitions.IsSelectionInGemocLanguageProject">
+                        </reference>
+                     </iterate>
+                  </with>
+               </visibleWhen>
+            </command>
+            <command
+                  commandId="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateXtextEditorProject"
+                  icon="icons/plus-16.png"
+                  label="Create XText Editor Project for language"
+                  style="push">
+               <visibleWhen
+                     checkEnabled="false">
+                  <with
+                        variable="selection">
+                     <count
+                           value="1">
+                     </count>
+                     <iterate
+                           ifEmpty="false">
+                        <reference
+                              definitionId="org.eclipse.gemoc.xdsmlframework.ide.ui.definitions.IsSelectionInGemocLanguageProject">
+                        </reference>
+                     </iterate>
+                  </with>
+               </visibleWhen>
+            </command>
+            <command
+                  commandId="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateAnimatorProject"
+                  icon="icons/plus-16.png"
+                  label="Create Animator Project for language"
+                  style="push">
+               <visibleWhen
+                     checkEnabled="false">
+                  <with
+                        variable="selection">
+                     <count
+                           value="1">
+                     </count>
+                     <iterate
+                           ifEmpty="false">
+                        <reference
+                              definitionId="org.eclipse.gemoc.xdsmlframework.ide.ui.definitions.IsSelectionInGemocLanguageProject">
+                        </reference>
+                     </iterate>
+                  </with>
+               </visibleWhen>
+            </command>
+         </menu>
+      </menuContribution>
    </extension>
    <extension
          id="GemocLanguageProjectBuilder"
@@ -152,6 +245,43 @@
             id="org.eclipse.gemoc.xdsmlframework.ide.ui.addRemoveGemocLanguageProjectNature"
             name="Add/Remove GEMOC Language Project Nature">
       </command>
+      <command
+            categoryId="org.eclipse.gemoc.xdsmlframework.ide.ui.GemocLanguageProjectNatureCategory"
+            defaultHandler="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateDomainModelProjectHandler"
+            id="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateDomainModelProject"
+            name="CreateDomainModelProject">
+      </command>
+      <command
+            categoryId="org.eclipse.gemoc.xdsmlframework.ide.ui.GemocLanguageProjectNatureCategory"
+            defaultHandler="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateXtextEditorProjectHandler"
+            id="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateXtextEditorProject"
+            name="CreateXtextEditorProject">
+      </command>
+      <command
+            categoryId="org.eclipse.gemoc.xdsmlframework.ide.ui.GemocLanguageProjectNatureCategory"
+            defaultHandler="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateSiriusEditorProjectHandler"
+            id="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateSiriusEditorProject"
+            name="CreateSiriusEditorProject">
+      </command>
+      <command
+            categoryId="org.eclipse.gemoc.xdsmlframework.ide.ui.GemocLanguageProjectNatureCategory"
+            defaultHandler="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateAnimatorProjectHandler"
+            id="org.eclipse.gemoc.xdsmlframework.ide.ui.commands.CreateAnimatorProject"
+            name="CreateAnimatorProject">
+      </command>
+   </extension>
+   <extension
+         point="org.eclipse.core.expressions.definitions">
+      <definition
+            id="org.eclipse.gemoc.xdsmlframework.ide.ui.definitions.IsSelectionInGemocLanguageProject">
+         <adapt
+               type="org.eclipse.core.resources.IResource">
+            <test
+                  property="org.eclipse.core.resources.projectNature "
+                  value="org.eclipse.gemoc.xdsmlframework.ide.ui.GemocLanguageProjectNature">
+            </test>
+         </adapt>
+      </definition>
    </extension>
 
 </plugin>
